+++ X10D/src/Numerics/Vector4Extensions.cs
@@ -10,6 +10,22 @@ namespace X10D.Numerics;
 public static class Vector4Extensions
 {
     /// <summary>
+    ///     Deconstructs the current <see cref="Vector4" /> into its components.
+    /// </summary>
+    /// <param name="vector">The vector to deconstruct.</param>
+    /// <param name="x">The X component value.</param>
+    /// <param name="y">The Y component value.</param>
+    /// <param name="z">The Z component value.</param>
+    /// <param name="w">The W component value.</param>
+    public static void Deconstruct(this Vector4 vector, out float x, out float y, out float z, out float w)
+    {
+        x = vector.X;
+        y = vector.Y;
+        z = vector.Z;
+        w = vector.W;
+    }
+
+    /// <summary>
     ///     Returns a vector whose Y, Z, and W components are the same as the specified vector, and whose X component is a new
     ///     value.
     /// </summary>